refactor(core): split FilterSheet into FilterSection and FilterSheetHeader files
Follows the one-component-per-file convention used across core/presentation/component (Loader, ErrorLayout, EmptySearch).
